Refactoring, no more 3-second delay in responses
Things are moved into a bot class. Instead of a giant 3-second delay,
the bot queues a message when it wants to reply, and the message is sent
within a second on another thread. No more sleeps on receiving messages,
which means bots should be much easier to quiet down. !rate 0 will be
received instantly.